team4/pyproject.toml
[project]
name = "pic-to-ascii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=8.32.0",
    "pillow>=11.1.0",
]
